basic "hello world" demo implemented
I created a couple of functions used to create rigid bodies, and set gravity. The PhysicsFacade class still needs to be implemented in a Singleton design pattern. I also implemented a std::list of btRigidBody to store the rigid bodies that are created, as the dynamicWorld variable doesn't provide easy access to the objects once added to it.