Updated the master branch
I have implemented all my physics changes and various other code additions in this push. I made small changes to several classes but the end result is that physics collisions are just about sorted. Still needs some polish but at the moment when the program is run one of the asteroids is static so it cannot be moved. This is just to demonstrate that the rigid bodies are there.