Added my changes from my PhysicsBranch
I changed some of the functions within PhysicsFacade, and also changed the asteroid and astroboy scripts so that they should work in our environment. This build compiles and runs, but I was unable to properly test the terrain and script changes due to the build being setup in a different way to test animation.