Error in GameManager.Lua file
The build will compile, but during runtime it throughs an exception regarding missing Lua5.1.4 symbols. I didn't touch any of the lua stuff other than trying to uncomment lines in GameManager.lua to bring models into the scene. The chunk of text in GameManager.lua, between lines 35 and 52 seem to be causing the issue.