Made minor changes to master
Made small changes to master, not ideal, but these changes don't change much so shouldn't impact too much. Builds and runs on my machine.
All I did was let the gme loop through the update method in gamecontroller once before triggering the menu to be displayed. This essentiallly results in the game starting with the menu being displayed.